+++ packages/dso-toolkit/src/components/badge/badge.models.ts
@@ -1,4 +1,4 @@
 export interface Badge {
-  status?: "primary" | "success" | "info" | "warning" | "danger" | "error" | "outline";
+  status?: "primary" | "success" | "info" | "warning" | "danger" | "error" | "outline" | "attention";
   message: string;
 }