+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## NEXT
 
+### Added
+* Extra kleurvariant `attention` voor label en badge ([#2008](https://github.com/dso-toolkit/dso-toolkit/issues/2008))
+
 ## 51.2.0
 
 ### Added